Adiciona classe ArredondarParaMeio e atualiza modelos e serviços

+++ POO/StreamingMovie/src/main/java/com/StreamingMovie/utils/Classificacao.java
@@ -2,5 +2,5 @@ package utils;
 
 public interface Classificacao {
 
-    int getClassificacao();
+    double getClassificacao();
 }
